requirements.txt
beautifulsoup4==4.13.4
brotli==1.1.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cssselect2==0.8.0
fonttools==4.57.0
google-api-core==2.24.2
google-api-python-client==2.118.0
google-auth==2.38.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.69.2
html5lib==1.1
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
pillow==11.1.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydyf==0.11.0
pyparsing==3.2.3
pypdf2==3.0.1
pyphen==0.17.2
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
soupsieve==2.7
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-extensions==4.13.2
uritemplate==4.1.1
urllib3==2.3.0
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
